@@ -4,19 +4,19 @@ description: JSON Schema implementation in Dart
 homepage: https://github.com/workiva/json_schema
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  collection: ^1.14.6
-  http: ">=0.12.0 <0.14.0"
-  logging: '>=0.9.3 <2.0.0'
-  rfc_6901: ^0.1.0
-  uri: '>=0.11.1 <2.0.0'
+  collection: ^1.16.0
+  http: ^0.13.4
+  logging: ^1.0.2
+  rfc_6901: ^0.1.1
+  uri: ^1.0.0
 
 dev_dependencies:
-  dependency_validator: ^3.1.2
-  path: ^1.3.0
-  shelf: ^0.7.9
-  shelf_static: ^0.2.9
-  test: ^1.9.1
+  dependency_validator: ^3.2.0
+  path: ^1.8.1
+  shelf: ^1.2.0
+  shelf_static: ^1.1.0
+  test: ^1.21.1
   workiva_analysis_options: ^1.2.2
