@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'dart:convert' as convert;
@@ -18,13 +19,18 @@ class IoSchemaUrlClient extends SchemaUrlClient {
   @override
   createFromUrl(
     String schemaUrl, {
-    SchemaVersion schemaVersion,
-    List<CustomVocabulary> customVocabularies,
-    Map<String, ValidationContext Function(ValidationContext context, String instanceData)> customFormats = const {},
+    SchemaVersion? schemaVersion,
+    List<CustomVocabulary>? customVocabularies,
+    Map<
+            String,
+            ValidationContext Function(
+                ValidationContext context, String? instanceData)>
+        customFormats = const {},
   }) async {
     final uriWithFrag = Uri.parse(schemaUrl);
-    final uri = schemaUrl.endsWith('#') ? uriWithFrag : uriWithFrag.removeFragment();
-    Map schemaMap;
+    final uri =
+        schemaUrl.endsWith('#') ? uriWithFrag : uriWithFrag.removeFragment();
+    Map? schemaMap;
     if (uri.scheme == 'http' || uri.scheme == 'https') {
       // Setup the HTTP request.
       _logger.info('GET\'ing Schema from URL: $uri');
@@ -39,10 +45,13 @@ class IoSchemaUrlClient extends SchemaUrlClient {
       final schemaText = await convert.Utf8Decoder().bind(response).join();
       schemaMap = json.decode(schemaText);
     } else if (uri.scheme == 'file' || uri.scheme == '') {
-      final fileString = await File(uri.scheme == 'file' ? uri.toFilePath() : schemaUrl).readAsString();
+      final fileString =
+          await File(uri.scheme == 'file' ? uri.toFilePath() : schemaUrl)
+              .readAsString();
       schemaMap = json.decode(fileString);
     } else {
-      throw FormatException('Url schema must be http, file, or empty: $schemaUrl');
+      throw FormatException(
+          'Url schema must be http, file, or empty: $schemaUrl');
     }
     // HTTP servers / file systems ignore fragments, so resolve a sub-map if a fragment was specified.
     final parentSchema = await JsonSchema.createAsync(
@@ -52,15 +61,17 @@ class IoSchemaUrlClient extends SchemaUrlClient {
       customVocabularies: customVocabularies,
       customFormats: customFormats,
     );
-    final schema = JsonSchemaUtils.getSubMapFromFragment(parentSchema, uriWithFrag);
+    final schema =
+        JsonSchemaUtils.getSubMapFromFragment(parentSchema, uriWithFrag);
     return schema ?? parentSchema;
   }
 
   @override
-  Future<Map<String, dynamic>> getSchemaJsonFromUrl(String schemaUrl) async {
+  Future<Map<String, dynamic>?> getSchemaJsonFromUrl(String schemaUrl) async {
     final uriWithFrag = Uri.parse(schemaUrl);
-    final uri = schemaUrl.endsWith('#') ? uriWithFrag : uriWithFrag.removeFragment();
-    Map schemaMap;
+    final uri =
+        schemaUrl.endsWith('#') ? uriWithFrag : uriWithFrag.removeFragment();
+    Map? schemaMap;
     if (uri.scheme == 'http' || uri.scheme == 'https') {
       // Setup the HTTP request.
       _logger.info('GET\'ing Schema JSON from URL: $uri');
@@ -75,10 +86,13 @@ class IoSchemaUrlClient extends SchemaUrlClient {
       final schemaText = await convert.Utf8Decoder().bind(response).join();
       schemaMap = json.decode(schemaText);
     } else if (uri.scheme == 'file' || uri.scheme == '') {
-      final fileString = await File(uri.scheme == 'file' ? uri.toFilePath() : schemaUrl).readAsString();
+      final fileString =
+          await File(uri.scheme == 'file' ? uri.toFilePath() : schemaUrl)
+              .readAsString();
       schemaMap = json.decode(fileString);
     } else {
-      throw FormatException('Url schema must be http, file, or empty: $schemaUrl');
+      throw FormatException(
+          'Url schema must be http, file, or empty: $schemaUrl');
     }
     // HTTP servers ignore fragments, so resolve a sub-map if a fragment was specified.
     var subSchema;
@@ -87,7 +101,7 @@ class IoSchemaUrlClient extends SchemaUrlClient {
     } catch (_) {
       // Do nothing if we fail to decode or read the pointer.
     }
-    return subSchema ?? schemaMap;
+    return subSchema ?? schemaMap as FutureOr<Map<String, dynamic>?>;
   }
 }
 
