@@ -38,8 +38,8 @@
 
 import 'package:json_schema/json_schema.dart';
 
-ValidationContext screamingCapsValidator(ValidationContext context, String instanceData) {
-  if (instanceData.toUpperCase() != instanceData) {
+ValidationContext screamingCapsValidator(ValidationContext context, String? instanceData) {
+  if (instanceData!.toUpperCase() != instanceData) {
     context.addError('"screaming-caps" format not accepted $instanceData');
   }
   return context;
@@ -58,6 +58,6 @@ main() {
     customFormats: customFormats,
   );
 
-  print(schema.validate({'baz': 'DO IT NOW'}, validateFormats: true).isValid);
-  print(schema.validate({'baz': 'do it'}, validateFormats: true).isValid);
+  print(schema.validate({'baz': 'DO IT NOW'}, validateFormats: true)!.isValid);
+  print(schema.validate({'baz': 'do it'}, validateFormats: true)!.isValid);
 }
