@@ -2,9 +2,9 @@ import 'package:json_schema/src/json_schema/global_platform_functions.dart';
 import 'package:json_schema/src/json_schema/models/schema_version.dart';
 import 'package:json_schema/src/json_schema/models/validation_context.dart';
 
-ValidationContext defaultUriReferenceValidator(ValidationContext context, String instanceData) {
+ValidationContext defaultUriReferenceValidator(ValidationContext context, String? instanceData) {
   if (context.schemaVersion < SchemaVersion.draft6) return context;
-  final isValid = defaultValidators.uriReferenceValidator ?? (_) => false;
+  final isValid = defaultValidators.uriReferenceValidator as dynamic Function(String?)? ?? (_) => false;
 
   if (!isValid(instanceData)) {
     context.addError('"uri-reference" format not accepted $instanceData');
