@@ -3,6 +3,7 @@
 // (one up from here). The `gen-fixtures` command in the Makefile will do the right thing.
 
 import 'dart:io';
+import 'package:collection/collection.dart' show IterableExtension;
 
 final String remotesDirectory = 'test/JSON-Schema-Test-Suite/remotes';
 final String remotesOutputFile = 'test/unit/specification_remotes.dart';
@@ -28,7 +29,8 @@ bool generateFileMapFromDirectory(
   Map<String, String> schemaFiles = {};
   for (var directory in directories) {
     for (var file in _jsonFiles(directory)) {
-      schemaFiles[file.path.replaceFirst(directory, host)] = file.readAsStringSync();
+      schemaFiles[file.path.replaceFirst(directory, host)] =
+          file.readAsStringSync();
     }
   }
 
@@ -52,7 +54,10 @@ String _generateRemoteEntries(Map<String, String> schemaFiles,
   final List<String> entries = [];
   bool isFirst = true;
   schemaFiles.forEach((String path, String fileContents) {
-    if (skipFiles.firstWhere((fileName) => path.endsWith(fileName), orElse: () => null)?.isNotEmpty == true) {
+    if (skipFiles
+            .firstWhereOrNull((fileName) => path.endsWith(fileName))
+            ?.isNotEmpty ==
+        true) {
       return;
     }
     entries.add(
@@ -75,12 +80,13 @@ List<File> _jsonFiles(String rootDir) {
   return files..sort((f1, f2) => f1.path.compareTo(f2.path));
 }
 
-void main([List<String> args]) {
+void main([List<String> args = const []]) {
   final argsMutableCopy = List.from(args);
   bool shouldCheck = argsMutableCopy.contains('--check');
   argsMutableCopy.removeWhere((arg) => arg == '--check');
 
-  if (argsMutableCopy.isEmpty || argsMutableCopy[0] == 'specification_remotes') {
+  if (argsMutableCopy.isEmpty ||
+      argsMutableCopy[0] == 'specification_remotes') {
     final bool didChange = generateFileMapFromDirectory(
       [remotesDirectory],
       remotesOutputFile,
@@ -88,7 +94,8 @@ void main([List<String> args]) {
       variableName: 'specificationRemotes',
     );
     if (didChange) {
-      print('Generation changed specification_remotes.dart. Please check in latest changes.');
+      print(
+          'Generation changed specification_remotes.dart. Please check in latest changes.');
       if (shouldCheck) exit(1);
     } else {
       print('No formatting changes were made for specification_remotes.');
@@ -103,7 +110,8 @@ void main([List<String> args]) {
       variableName: 'additionalRemotes',
     );
     if (didChange) {
-      print('Generation changed additional_remotes.dart. Please check in latest changes.');
+      print(
+          'Generation changed additional_remotes.dart. Please check in latest changes.');
       if (shouldCheck) exit(1);
     } else {
       print('No formatting changes were made for additional_remotes.');
@@ -118,7 +126,8 @@ void main([List<String> args]) {
       variableName: 'specificationTests',
     );
     if (didChange) {
-      print('Generation changed additional_remotes.dart. Please check in latest changes.');
+      print(
+          'Generation changed additional_remotes.dart. Please check in latest changes.');
       if (shouldCheck) exit(1);
     } else {
       print('No formatting changes were made for additional_remotes.');
